load new images for previews, calculate real height from aspect
