+++ convex/_generated/api.d.ts
@@ -18,7 +18,6 @@ import type * as features from "../features.js";
 import type * as imageMigration from "../imageMigration.js";
 import type * as internal_ from "../internal.js";
 import type * as migrations from "../migrations.js";
-import type * as previews from "../previews.js";
 import type * as projects from "../projects.js";
 
 /**
@@ -35,7 +34,6 @@ declare const fullApi: ApiFromModules<{
   imageMigration: typeof imageMigration;
   internal: typeof internal_;
   migrations: typeof migrations;
-  previews: typeof previews;
   projects: typeof projects;
 }>;
 export declare const api: FilterApi<
